change: - redo alarm facitlity:   * logic moved to MBean   * 1.0 and 1.1 are present in MBean,   * cleaned impl  - for some reason it used seaprate mechanism than provided by jboss to disspatch notifications - alarm facitlity is  created per sbb component upon creation, per ra entty (profiles todo)
git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/jain-slee@5025 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ core/jar/src/main/java/org/mobicents/slee/runtime/facilities/MNotificationSource.java
@@ -1,13 +1,14 @@
 package org.mobicents.slee.runtime.facilities;
 
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import javax.slee.management.NotificationSource;
 
 public class MNotificationSource {
 
 	private NotificationSource notificationSource = null;
-	private AtomicInteger sequence = new AtomicInteger(0);
+	private AtomicLong sequence = new AtomicLong(0);
 	
 	
 	
@@ -45,7 +46,7 @@ public class MNotificationSource {
 	public NotificationSource getNotificationSource() {
 		return notificationSource;
 	}
-	public int getNextSequence() {
+	public long getNextSequence() {
 		return sequence.getAndIncrement();
 	}
